requirements.txt
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
httpx==0.27.2
yfinance==0.2.50
pandas==2.2.3
numpy==2.1.3
prometheus-client==0.21.1
praw==7.8.1
openai==1.55.3
mcp==1.1.2
rich==13.9.4
click==8.1.7
